sdk/reader: split claude relationship inference into a submodule
Extract the ~520-line session-relationship cluster (explicit fork/
continuation rows, inferred parent-UUID continuations and shared-source
forks, /resume markers, subagent-spawn and compaction-event annotation,
and cross-file reconciliation) out of the 5.6k-line claude.rs into
claude/relationships.rs. Pure move: shared low-level accessors
(string_field, extract_plain_user_text_from_obj, *_KEYS consts) are now
pub(super); the parse engine drives the helpers via re-exports. Public
surface (reconcile_claude_session_relationships, ClaudeRelationshipEvidence,
ReconcileClaudeRelationshipsInput) unchanged.
